+++ docs/conf.py
@@ -24,6 +24,8 @@ author = 'Electro707'
 # The full version, including alpha/beta/rc tags
 release = __version__
 
+autodoc_member_order = 'bysource'
+
 
 # -- General configuration ---------------------------------------------------
 
@@ -33,7 +35,8 @@ release = __version__
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.todo",
-    "sphinx.ext.inheritance_diagram"
+    "sphinx.ext.inheritance_diagram",
+    "sphinx.ext.napoleon"
 ]
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']